1. Changed loop logic again 2. Added test to handle off by one errors in coding grad 3. Moved tests into grad_test.cpp. (Renaming from old agrad_test.cpp and removing grad_nested_rev_test.cpp)
(Issue #1855)
